Fixed 2 Bugs
Bug 1: White space was counting as a click. Modified the qwerty event listener to check for the tag name being button and checking for NOT including the chosen class separately.

Bug 2: If a game is lost, then the next game won, the background of the overlay stays red and vice versus. Was abkle to fix by resetting the #overlay classes in the mainButton resets.